+++ CHANGELOG.md
@@ -36,6 +36,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## Unreleased
+
+### Bug Fixes
+
+* (server) [ENG-885](https://linear.app/liftedinit/issue/ENG-885) Reconstruct in-place testnet commits with vote extensions and update the cached genesis chain ID.
+
 ## [v0.50.14](https://github.com/cosmos/cosmos-sdk/releases/tag/v0.50.14) - 2025-07-08
 
 ### Bug Fixes